+++ payment_lyra/models/payment_provider.py
@@ -12,15 +12,15 @@ from datetime import datetime
 from hashlib import sha1, sha256
 import hmac
 import logging
-import math
 from os import path
 
 from pkg_resources import parse_version
 
-from odoo import models, api, release, fields, _
+from odoo import models, api, fields, _
 from odoo.exceptions import ValidationError
 from odoo.tools import convert_xml_import
 from odoo.tools import float_round
+from odoo.tools import get_lang
 from odoo.tools.float_utils import float_compare
 from odoo.http import request
 
@@ -37,6 +37,7 @@ _logger = logging.getLogger(__name__)
 
 class ProviderLyra(models.Model):
     _inherit = 'payment.provider'
+    _name = 'payment.provider'
 
     def _get_notify_url(self):
         base_url = self.env['ir.config_parameter'].get_param('web.base.url')
@@ -59,6 +60,22 @@ class ProviderLyra(models.Model):
 
         return doc_field_html
 
+    def _get_payment_data_entry_mode(self):
+        payment_data_entry_mode = constants.LYRA_PAYMENT_DATA_ENTRY_MODE
+        if constants.LYRA_PLUGIN_FEATURES.get('smartform') == False:
+            del payment_data_entry_mode["smartform"]
+            del payment_data_entry_mode["smartform_extended_with_logos"]
+            del payment_data_entry_mode["smartform_extended_without_logos"]
+
+        return [(c, _(l)) for c, l in payment_data_entry_mode.items()]
+
+    def _get_default_entry_mode(self):
+        module_upgrade = request.env['ir.module.module'].search([('state', '=', 'to upgrade'), ('name', '=', 'payment_lyra')])
+        if module_upgrade or (constants.LYRA_PLUGIN_FEATURES.get('smartform') == False):
+            return ("redirect")
+
+        return ("smartform")
+
     sign_algo_help = _('Algorithm used to compute the payment form signature. Selected algorithm must be the same as one configured in the Lyra Expert Back Office.')
 
     if constants.LYRA_PLUGIN_FEATURES.get('shatwo') == False:
@@ -79,11 +96,10 @@ class ProviderLyra(models.Model):
     lyra_key_prod = fields.Char(string=_('Key in production mode'), help=_('Key provided by Lyra Collect (available in Lyra Expert Back Office after enabling production mode).'), default=constants.LYRA_PARAMS.get('KEY_PROD'))
     lyra_sign_algo = fields.Selection(string=_('Signature algorithm'), help=sign_algo_help, selection=[('SHA-1', 'SHA-1'), ('SHA-256', 'HMAC-SHA-256')], default=constants.LYRA_PARAMS.get('SIGN_ALGO'))
     lyra_notify_url = fields.Char(string=_('Instant Payment Notification URL'), help=_('URL to copy into your Lyra Expert Back Office > Settings > Notification rules.'), default=_get_notify_url, readonly=True)
-    lyra_gateway_url = fields.Char(string=_('Payment page URL'), help=_('Link to the payment page.'), default=constants.LYRA_PARAMS.get('GATEWAY_URL'))
     lyra_language = fields.Selection(string=_('Default language'), help=_('Default language on the payment page.'), default=constants.LYRA_PARAMS.get('LANGUAGE'), selection=_get_languages)
     lyra_available_languages = fields.Many2many('lyra.language', string=_('Available languages'), column1='code', column2='label', help=_('Languages available on the payment page. If you do not select any, all the supported languages will be available.'))
-    lyra_capture_delay = fields.Char(string=_('Capture delay'), help=_('The number of days before the bank capture (adjustable in your Lyra Expert Back Office).'))
-    lyra_validation_mode = fields.Selection(string=_('Validation mode'), help=_('If manual is selected, you will have to confirm payments manually in your Lyra Expert Back Office.'), selection=[('-1', _('Lyra Expert Back Office Configuration')), ('0', _('Automatic')), ('1', _('Manual'))])
+    lyra_capture_delay = fields.Char(string=_('Capture delay (if applicable)'), help=_('The number of days before the bank capture (adjustable in your Lyra Expert Back Office).'))
+    lyra_validation_mode = fields.Selection(string=_('Validation mode (if applicable)'), help=_('If manual is selected, you will have to confirm payments manually in your Lyra Expert Back Office.'), selection=[('-1', _('Lyra Expert Back Office Configuration')), ('0', _('Automatic')), ('1', _('Manual'))])
     lyra_payment_cards = fields.Many2many('lyra.card', string=_('Card types'), column1='code', column2='label', help=_('The card type(s) that can be used for the payment. Select none to use gateway configuration.'))
     lyra_threeds_min_amount = fields.Char(string=_('Manage 3DS'), help=_('Amount below which customer could be exempt from strong authentication. Needs subscription to «Selective 3DS1» or «Frictionless 3DS2» options. For more information, refer to the module documentation.'))
     lyra_redirect_enabled = fields.Selection(string=_('Automatic redirection'), help=_('If enabled, the buyer is automatically redirected to your site at the end of the payment.'), selection=[('0', _('Disabled')), ('1', _('Enabled'))])
@@ -98,9 +114,21 @@ class ProviderLyra(models.Model):
     lyra_multi_period = fields.Char(string=_('Period'), help=_('Delay (in days) between installments.'))
     lyra_multi_first = fields.Char(string=_('1st installment'), help=_('Amount of first installment, in percentage of total amount. If empty, all installments will have the same amount.'))
 
-    image = fields.Char()
-    environment = fields.Char()
+    lyra_test_password = fields.Char(string=_('Test password'), help=_(constants.LYRA_REST_API_KEYS_DESC))
+    lyra_prod_password = fields.Char(string=_('Production password'), help=_(constants.LYRA_REST_API_KEYS_DESC))
+    lyra_public_test_key = fields.Char(string=_('Public test key'), help=_(constants.LYRA_REST_API_KEYS_DESC))
+    lyra_public_production_key = fields.Char(string=_('Public production key'), help=_(constants.LYRA_REST_API_KEYS_DESC))
+    lyra_sha256_test_key = fields.Char(string=_('HMAC-SHA-256 test key'), help=_(constants.LYRA_REST_API_KEYS_DESC))
+    lyra_sha256_prod_key = fields.Char(string=_('HMAC-SHA-256 production key'), help=_(constants.LYRA_REST_API_KEYS_DESC))
+    lyra_rest_api_notify_url = fields.Char(string=_('REST API Notification URL'), help=_('URL to copy into your Lyra Expert Back Office > Settings > Notification rules.'), default=_get_notify_url, readonly=True)
 
+    lyra_payment_data_entry_mode = fields.Selection(string=_('Payment data entry mode'), help=_('Select how the payment data will be entered. Attention, to use the Smartform, you must ensure that you have subscribed to this option with Lyra Collect.'), selection=_get_payment_data_entry_mode, default=_get_default_entry_mode, readonly=not(constants.LYRA_PLUGIN_FEATURES.get('smartform')))
+    lyra_smartform_pop_in = fields.Selection(string=_('Display in a pop-in'), help=_('This option allows to display the Smartform in a pop-in.'), selection=[('no', _('No')), ('yes', _('Yes'))], default='no')
+    lyra_smartform_theme = fields.Selection(string=_('Theme'), help=_('Select a theme to use to display the Smartform.'), selection=[('neon', 'Neon'), ('classic', 'Classic')], default='neon')
+    lyra_smartform_compact_mode = fields.Selection(string=_('Compact mode'), help=_('This option allows to display the Smartform in a compact mode.'), selection=[('0', _('Disabled')), ('1', _('Enabled'))], default='0')
+    lyra_smartform_payment_attempts = fields.Char(string=_('Payment attempts number for cards'), help=_('Maximum number of payment by cards retries after a failed payment (between 0 and 2). If blank, the gateway default value is 2.'))
+
+    environment = fields.Char()
     lyra_redirect = False
 
     @api.model
@@ -117,16 +145,15 @@ class ProviderLyra(models.Model):
         return providers
 
     @api.model
-    def multi_add(self, filename):
-        if (constants.LYRA_PLUGIN_FEATURES.get('multi') == True):
+    def multi_add(self, filename, noupdate):
+        if constants.LYRA_PLUGIN_FEATURES.get('multi') == True:
+            module_upgrade = request.env['ir.module.module'].search([('state', '=', 'to upgrade'), ('name', '=', 'payment_lyra')])
             file = path.join(path.dirname(path.dirname(path.abspath(__file__)))) + filename
-            convert_xml_import(self._cr, 'payment_lyra', file)
-
-        return None
+            mode = 'update' if module_upgrade else 'init'
+            convert_xml_import(self._cr, 'payment_lyra', file, None, mode, noupdate)
 
     def _get_ctx_mode(self):
-        ctx_key = self.state
-        ctx_value = 'TEST' if ctx_key == 'test' else 'PRODUCTION'
+        ctx_value = 'TEST' if self.state == 'test' else 'PRODUCTION'
 
         return ctx_value
 
@@ -149,7 +176,7 @@ class ProviderLyra(models.Model):
 
     def _get_payment_config(self, amount):
         if self.code == 'lyramulti':
-            if (self.lyra_multi_first):
+            if self.lyra_multi_first:
                 first = int(float(self.lyra_multi_first) / 100 * int(amount))
             else:
                 first = int(float(amount) / float(self.lyra_multi_count))
@@ -221,7 +248,7 @@ class ProviderLyra(models.Model):
             'vads_url_return': urlparse.urljoin(base_url, LyraController._return_url),
             'vads_order_id': str(order_id),
             'vads_ext_info_order_ref': str(values.get('reference')),
-            'vads_contrib': constants.LYRA_PARAMS.get('CMS_IDENTIFIER') + u'_' + constants.LYRA_PARAMS.get('PLUGIN_VERSION') + u'/' + release.version,
+            'vads_contrib': tools._lyra_get_contrib(),
 
             'vads_language': self.lyra_language or '',
             'vads_available_languages': available_languages,
@@ -247,11 +274,7 @@ class ProviderLyra(models.Model):
                 tx_values[key] = ''
 
             lyra_tx_values[key] = tx_values[key].encode('utf-8')
-
         return lyra_tx_values
-
-    def lyra_get_form_action_url(self):
-        return self.lyra_gateway_url
 
     def _get_default_payment_method_id(self, code):
         self.ensure_one()
@@ -262,3 +285,64 @@ class ProviderLyra(models.Model):
             return self.env.ref('payment_lyra.payment_method_lyra').id
         if self.code == 'lyramulti':
             return self.env.ref('payment_lyra.payment_method_lyramulti').id
+
+    def _lyra_get_rest_password(self):
+        if self.state == 'test':
+            return str(self.lyra_test_password)
+
+        return str(self.lyra_prod_password)
+
+    def _lyra_get_rest_public_key(self):
+        if self.state == 'test':
+            return str(self.lyra_public_test_key)
+
+        return str(self.lyra_public_production_key)
+
+    def _lyra_get_rest_sha256_key(self):
+        if self.state == 'test':
+            return str(self.lyra_sha256_test_key)
+
+        return str(self.lyra_sha256_prod_key)
+
+    def _lyra_get_javascript_server_url(self):
+        return constants.LYRA_PARAMS.get('STATIC_URL') + "js/krypton-client/V4.0/stable/kr-payment-form.min.js"
+
+    def _lyra_get_smartform_stylesheet_url(self):
+        return constants.LYRA_PARAMS.get('STATIC_URL') + "js/krypton-client/V4.0/ext/" + self.lyra_smartform_theme + "-reset.css"
+
+    def _lyra_get_smartform_stylesheet_script_url(self):
+        return constants.LYRA_PARAMS.get('STATIC_URL') + "js/krypton-client/V4.0/ext/" + self.lyra_smartform_theme + ".js"
+
+    def _lyra_get_return_url(self):
+        return urlparse.urljoin(self.env['ir.config_parameter'].get_param('web.base.url'), LyraController._return_url)
+
+    def _lyra_get_smartform_language(self):
+        return get_lang(self.env).code[:2]
+
+    def _lyra_get_smartform_payment_means(self):
+        cards = []
+        for value in self.lyra_payment_cards:
+            cards.append(value.code)
+
+        return cards
+
+    def _lyra_get_currency(self, currency_id):
+        # Give the iso and the number of decimal toward the smallest monetary unit from the id of the currency.
+        currency_number = tools.find_currency(self.env['res.currency'].browse(currency_id).exists().name)
+        for currency in constants.LYRA_CURRENCIES:
+            if currency[1] == str(currency_number):
+                return (currency[0], currency[2])
+
+        raise ValidationError(_('The shop currency {} is not supported.').format(currency.name))
+
+    def _should_build_inline_form(self, is_validation=True):
+        if (self.code == 'lyramulti'):
+            return False
+
+        if (self.code == 'lyra'):
+            if (constants.LYRA_PLUGIN_FEATURES.get('smartform') == True and not str(self.lyra_payment_data_entry_mode) == 'redirect'):
+                return True
+
+            return False
+
+        return super()._should_build_inline_form(is_validation=True)